+++ watch/process_group_windows.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package watch
+
+import (
+	"os"
+	"os/exec"
+)
+
+func configureProcessGroup(_ *exec.Cmd) {}
+
+func signalProcessGroup(process *os.Process, sig os.Signal) error {
+	return process.Signal(sig)
+}